Move teleport manager into a new API.
Hopefully, this will help me get Logistics Pipes compatibility.
