+++ src/main/java/buildcraft/additionalpipes/api/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Additional Pipes API - just for teleport pipes at this point.
+ */
+@API(apiVersion = "1", owner = "additionalpipes", provides = "apAPI")
+package buildcraft.additionalpipes.api;
+
+import cpw.mods.fml.common.API;
+
+
