+++ src/main/java/buildcraft/additionalpipes/gates/GateProvider.java
@@ -6,8 +6,8 @@ import java.util.LinkedList;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraftforge.common.util.ForgeDirection;
 import buildcraft.additionalpipes.AdditionalPipes;
+import buildcraft.additionalpipes.api.PipeTeleport;
 import buildcraft.additionalpipes.pipes.PipeItemsClosed;
-import buildcraft.additionalpipes.pipes.PipeTeleport;
 import buildcraft.api.statements.IStatementContainer;
 import buildcraft.api.statements.ITriggerExternal;
 import buildcraft.api.statements.ITriggerInternal;
