+++ src/main/java/buildcraft/additionalpipes/gui/GuiHandler.java
@@ -3,8 +3,8 @@ package buildcraft.additionalpipes.gui;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.world.World;
+import buildcraft.additionalpipes.api.PipeTeleport;
 import buildcraft.additionalpipes.pipes.PipeItemsJeweled;
-import buildcraft.additionalpipes.pipes.PipeTeleport;
 import buildcraft.additionalpipes.pipes.PipeTransportAdvancedWood;
 import buildcraft.additionalpipes.utils.Log;
 import buildcraft.transport.TileGenericPipe;
