+++ src/main/java/buildcraft/additionalpipes/gui/GuiTeleportPipe.java
@@ -5,9 +5,9 @@ import net.minecraft.client.gui.GuiButton;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.IInventory;
 import buildcraft.BuildCraftCore;
+import buildcraft.additionalpipes.api.PipeTeleport;
 import buildcraft.additionalpipes.network.PacketHandler;
 import buildcraft.additionalpipes.network.message.MessageTelePipeUpdate;
-import buildcraft.additionalpipes.pipes.PipeTeleport;
 import buildcraft.additionalpipes.textures.Textures;
 import buildcraft.core.CoreIconProvider;
 import buildcraft.core.lib.gui.GuiBuildCraft;
